Implements Travis-CI Tests at gate
When merging this patch, it is also required to add the project
on the travis-ci UI. Simply create an account and it will pull in
repos to enable. Its a one click operation.

The `lukehinds` docker hub is used, but we can easily swap this out
for a project based account if prefered.

Closes issue #9