Docker changes required to replace tpm simulator
Previously ibmtpm974 was used. This change implements ibmtpm1119
inline with typical use and upstream (tpm-software/*)

This is achieved by updating the Dockerfile for TPM 2.0 and
uploading a new build with the tag `v301` to docker hub.

The travis.yml file is also repointed to the new tag.

test_wrapper.sh is updated to work with the new swtmp.