Allows tpm2-abrmd restart for test run
This change introduces a slight change to the package manager
selection and in turn installs python-dbus

Further changes are made to `test_restful.py` whereby if the
system is running within systemd mode with an active dbus, restart
`tpm2-abrmd.service`. If not, do not try to restart the tpm2-abrmnd.

The logic being, a non systemd based service will be running a
tpm1.2 system.

The service restart is required as when using mssim, the resource
manager cannot reconnect.

This allows continued use of one set of test scripts for both
TPM versions.