Move installer.sh to using packages
This change moves the installer to using OS packages, instead
of compiling.

It also adds logic to handle tpm2-abrmd working with an emulator
and provides a message on selinux (if a rhel based OS). This
message is to ask for SELinux to be disabled, as it blocks the
emulator and tpm2-abrmd from communication. Note SELinux disabling
is not required when tpm2-abrmd communincates with a hardware
TPM.

An emulator security warning is also displayed at the end of
installation about production use.