Set openssl as the default
This change switches the default to openssl.

The idea is that openssl is a rudimentary implementation that
works out of the box with no additional installation steps
required.

This is not to suggest in anyway that cfssl is a second class
citizen, just more that openssl is a better default for testing
and base installations. Should someone want to use CFSSL they
can then perform the installation afterwards.

Another benefit to openssl, is that Keylime can then be deployed
on an air-gapped machine with no need to retrieve anything from
the internet (everything else is in packages now so can be served
by a mirror)

Last of all #145 is the future and CA's will be a pluggable
framework