Fedora 31 based CI image
For some reason the systemd-udev package is not installed by
default in the base container image.
Because of that, we see this error when starting the tpm2-abrmd
service:
Failed to start tpm2-abrmd.service: Unit systemd-udev-settle.service not found.

Also commenting out starting condition that requires /dev/tmp* to
be present and not needed in CI.

Fix #271