Improve base http request
Make timeout configurable from argument, and we can make it
configurable in the configuration file in the future.

Remove retry count as it doesn't work as expected, we can add retry
mechanism after we sort out which kind of status should be retried.

Use the same code path whether there is body in the request because it's
only a difference in the argument.

Also renamed a few arguments to reflect their real meaning to be less
confusing.