Fixes incompatibility between keylime and tpm2-tools >= 4.2 (#273)
* Fixes incompatibility between keylime and tpm2-tools >= 4.2

This commit implements a fix for the keylime_agent when used with a more
recent (03/2020) version of tpm2-tools. At this point, this commit makes
the minimal amount of changes required to have keylime operational, by
simply defining a new "version point" (4.2) and adding the appropriate
conditionals in the multiple if-elif-else statements. Moving forward, a
cleaner refactoring (e.g. using a "command vs tpm2-tools version"
dictionary) looks like a more robust solution in the long term.

Resolves: #272

* Modified according to suggestion from Riccardo Pittau (elfosardo)

Co-authored-by: Marcio Silva <marcios@us.ibm.com>