Perform clean up and archive old items
The current `scripts` directory contains a lot of stuff we no
longer use, such as old test scripts and legacy vTPM scripts.

Lets archive this and use `scripts` for more utilised objects.

I also move create_whitelist.sh from keylime/keylime which is a
bit an odd home for it to reside.