Catch regex exception for invalid exclude list (#303)
This change checks that the regular expression provided in an exclude
list can be successfully compiled before attempting to match paths
against it. Prior to this change, if the exclude list contained an
invalid regular expression, an exception was raised when attempting to
match a path with the exclude list regular expression. This exception
was not successfully handled triggering the verifier to crash that could
result in a denial of service attack.