Agent can just "use" an already created EK (#320)
* Fix for registrar and verifier running on different hosts

This commit fixes the inability of running `keylime_tenant` in a setup
where Keylime's Registrar and Verifier daemons run on separate hosts.
There is also an attempt at providing a clearer error message when the
registrar cannot be reached at the specified IP address, TCP port.

Resolves: #297

* This commit is a first attempt at giving keylime de ability to use an
already generated Endorsement Key (EK) on a TPM. The use case and
rationale for this is well described on #313.

A new option is introduced under `[cloud_agent]`, name `ek_handle`. When
the default value of `generate` (wanted to keep the familiar term
already present on other options, such as `tls_dir`) is found, the
normal agent startup process, which includes clearing/start, ownership,
flush/evict and EK creation, will be executed. If `ek_handle` points to
an actual value (e.g., `0x81000000`), then all the initial steps
previously described will be skipped and the agent will just start from
the "create AIK" step. In the latter scenario, the agent will make sure
that particular handle is not flushed on exit.

Resolves: #313

Co-authored-by: Marcio Silva <marcios@us.ibm.com>