setup: Remove PBR once more (#338)
PBR is a very helpful tool. However, PBR is not available as a build
dependency in all cases. Here we switch back to the more traditional
in-line setup.py while attempting to honor the the content that
was in setup.cfg.

References:
- PBR added 1ae26fdabb30c54bbf08d8621c52e16020583c10
- Attempted removal cb42ba3b4295f8b0171209f70f1cf2513096150f
- Revert 401d4f9e75be361f91f9a12f004b633d36a0089e

* setup.py: Use ASL in trove classifiers

See https://pypi.org/pypi?%3Aaction=list_classifiers