ima: Add basis for file signature verification
This patch adds the basis for file signature verification. We
model a keyring where we can add public keys to that will be used
for signature verification. The keys are added to a map where the
indices are the key Id v2 (keyidv2) that is calculated from a
hash over the DER representation of the public key. RSA, DSA, and
elliptic curve keys are supported.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>