Use requests library instead of http.client
This means instead of passing around an SSL context to make requests we
just provide the paths to client-side tls cert and key.

This work was originally done by Luke Hinds (lhinds@redhat.com) as part
of his work on Multi-Tenancy. I've separated it out into a separate
commit since it's useful on it's own.