PR 1 (of 3) to address #491
* The `agent` will now retrieve TPM 2.0 preboot event log and will ship
it to the `verifier`. The latter will do nothing with it (just receives
and ignores it)
* Adds a new `keylime_tenant` CLI option, `--mb_refstate` (that was a
request from @mpeters) to signal the need for the agent to collect the
the event log. At this point, `keylime_tenant` do not actually process
this option (it shall be processed in a manner similar to IMA in a
future PR), it is simply used to signal the agent that "tpm2 preboot
event log" should be read.
* Hard-coded PCRs [0-9], 14 and path
/sys/kernel/security/tpm0/binary_bios_measurements for the event log
* The `verifier` will output a debug logging message stating that it
received the "tpm2 event log", and will send it down to the `check_pcrs`
method, but there is no actual processing there (it will be done in
subsequent PR)