ima: Reduce logging output on file signatures evaluation
Move the logger from info to debug to reduce logging output on
good file signatures.

Remove logging when a file signature cannot be verified. A log
entry will be emitted by the caller in ima.py.

Adjust all other logging levels that show that something is
not supported or the signature is invalid to 'error'.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>