ima: Error on path not in allowlist only if entry not evaluated by signature
This change does not change the behavior of the system if no IMA file
signature verification keys are provided, which should be the default
for now.

Tweak the evaluation of measurement entries so that if a file has
already been evaluated by a file signature then a missing path entry
in the allowlist is NOT an error. This allows us to only sign files
but does not require an allow list entry. However, if an allowlist entry
is available for a file, then it must match the IMA measurement.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>