SQLAlchemy database schema fix for the new `ima_sign_verification_keys` column
This commit, very similar to #376, fixes the SQLAlchemy problem that
appears when one attempts to create a column of type `String` with a
large value (`1048576`) against a MySQL database. This results in the
error `Column length too big for column 'ima_sign_verification_keys'
(max = 65535); use BLOB or TEXT instead")` and here we just change the
`ima_sign_verification_keys` column to `Text(429400000)`. Tested on both
MySQL 5.7.31 and SQLite.

Signed-off-by: Marcio Silva <marcios@us.ibm.com>