Second PR (of a total of 4) to address #491
* The most important change here is the fact that the TPM 2.0 preboot
event log , shipped to the verifier from the agent, will now be
authenticated, by replaying the aforementioned log and comparing it
against the values of PCRs [0-9],14.
* If these do not match, then the node will be considered as failed at
attestation.
* A small utility (tpm_bootlog_enrich.py) is added to perform the log replay.
In the future, it could be replaced by newer versions of tpm2-tools capable
of performing the same activity (several open PRs against the tool with
fixes and improvements)

Signed-off-by: Marcio Silva <marcios@us.ibm.com>