Consistent formatting for logging strings
Use standard python logging format where args are passed to the logging
functions (info(), error(), etc). Not only is this more consistent,
but it's better performant as interpolation doesn't happen until/if the
logs are written.

Before we were all over the place with fstrings, string concat, format
strings (but interpolation done before the function call).

Signed-off-by: Michael Peters <mpeters@redhat.com>