tenant: Support URLs and signatures for local and remote file signing keys
Extend the options parser with options for taking IMA file signing keys'
GPG signature files and GPG public keys for signature verification. Also
add support for remote IMA signature verification keys, remote GPG
signature files, and GPG public keys for verifying the remote signing
keys' signature.

The already-existing options parser supported reading of multiple keys
from files. With this extensions those key files can now be signed and
their signature can be verified. If signature verfication is supposed
to be done for a subset of the keys (local or remote) then those
have to be enumerated *first* and paired up with the signature file
option (local or remote) and the GPG public key to be used for
signature verification.

The following is an example:

keylime_tenant \
   [...] \
   --signature-verification-key /var/lib/keylime/imapubkeys/secp256k1.pem \
   --signature-verification-key-sig /var/lib/keylime/imapubkeys/secp256k1.pem.asc \
   --signature-verification-key-sig-key ./mygpg.pubkey  \
   \
   --signature-verification-key /var/lib/keylime/imapubkeys/rsa2048.pem  \
   --signature-verification-key ./ima-pub.pem \
   \
   --signature-verification-key-url http://remote.com/keys/rsa2048pub.pem \
   --signature-verification-key-sig-url http://remote.com/keys/rsa2048pub.pem.asc \
   --signature-verification-key-sig-url-key ./remotegpg.pubkey

The above shows a signed local key with its signature file (secp256k1.pem.asc)
and the GPG public key to use for signature verification (mygpg.pubkey). This
is followed by two local keys that are not signed. This again is followed by
a remote key (rsa2048pub.pem) and its signature file (rsa2048pub.pem.asc) and
the GPG public key to use for signature verification (remotegpg.pubkey).

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>