Explicitly test for tpm2_tools version 4.0 and 4.2 in else branch
There are several places in tpm_main.py where we explicitly test for 4.0 or 4.2
versions of the tpm2 tools in an else branch while testing for 3.2 in the if
branch. This patch adjusts the cases where we don't check for this in the else
branch explicitly while also checking for 3.2 in the if branch.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>