agent: Drop privileges by switching to normal user and group
Extend the agent section in keylime.conf with a runas parameter where
one can specify the user and group to run the agent as. We now have to
recursively change the ownership of the config.WORK_DIR so that once
we switched to the new user and group we can access the files and
directories there. After that we drop privileges to the specified user
and group, which is by default 'nobody:nobody'.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>