agent: Drop privileges by switching to normal user and group
Extend the agent section in keylime.conf with a runas parameter where
one can specify the user and group to run the agent as.

If the agent was started as root check whether run_as is specified and
if this is the case change the onwership on the tmpfs to the run_as
user and after that switch to the user.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>