config: Add optional fallback parameter to get()
Add an optional fallback parameter with the default value to the config's
get() function and related ones like getint(). If a fallback parameter is
given we will return that fallback parameter in case the section is missing
the given option. The variant without the fallback parameter will throw
an exception in case of a missing parameter.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>