ima_file_signatures: Extract keyidv2 from x509 certs
When an x509 certificate is given, attempt to extract the keyidv2 from the
certificate's subject key identifier digest's last 4 bytes. This is safer
than recalculating it with the sha1 hash since other methods may exist.
For public and public keys extracted from private keys we have to use the
sha1 method.

Also adjust how we store the state of the keys into the database. We now
also store an array of the keyidv2's of all keys.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>