installer: Fix back level Intel TSS configure issues for Ubuntu groovy
The first configure issue was reported here.  The recommendation
was to try 2.4.x, as 2.0.x is no longer supported.  This should
be tested against keylime 'next' builds for compatibility.

https://github.com/tpm2-software/tpm2-tss/issues/2066

The second configure issue was reported here.  The recommendation was
to disable building of the man pages, because 2.4.x expects al older
version of doxygen.

https://github.com/tpm2-software/tpm2-tss/issues/2067

Signed-off-by: Ken Goldman <kgold@linux.ibm.com>