tpm_abstract: Fix ioctl number for adding entropy on x86_64
The previous change fixed the invalid argument error for the ioctl to add
entropy to /dev/random on ppc64le but broke it on x86_64 since the ioctl
numbers are architecture-dependent. This patch fixes it now for x86_64 and
ppc64le and skips all other systems that have an unknown ioctl number.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>