agent: Add support for optional IMA log offset parameter
Add support for an optional IMA log entry parameter so that a newer
verifier can request that the IMA log be returned only from a given
entry. If the optional parameter is not given, we return the IMA log
from 0-th entry as before, thus supporting older verifier.

The verifier is only allowed to query the agent from a given entry.
The agent initializes this value to '0' at startup and moves it along
as the verifier receives result. This is to prevent that a verifer
queries as an agent at a higher entry number after restart of the
agent or reboot of the system the agent is on.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>