webapp: fix tls certs paths
`get_tls_context` search for the TLS certificates and TLS private key.
To do that uses a simple "%s/%s" to build the full path.

The problem is that if the certificates contains already the full path,
the final path will be incorrect.

This patch replace it with `os.path.join`, that respect the full path to
generate the final path.

Signed-off-by: Alberto Planas <aplanas@suse.com>