Introduce an AgentAttestState class for passing state through the APIs
To be able to support incremental attestation we need some sort of
tracking the state of the PCRs and how far along we received the IMA
measurement log so that we can continue the attestation where we left
off the last time. The AgentAttestState keeps track of the attestation
state of the system where each agent is located.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>