verifier: Extend AgentAttestState with state of the IMA PCR
To be able to deal with incremental IMA logs we must also model the
IMA PCR as part of the AgentAttestState and pass it now all the way into
process_measurement_list() as 1st parameter.

The cleanest way seems to be to model all the PCRs of the TPM in a class
of its own.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>