Consult the AgentAttestState for the next measurement list entry
Consult the AgentAttestState for the next measurement list entry to request
from the agent. If the agent returns us nothing or a 0, we reset the
attestation state, meaning that we expect to receive the whole measurement
list.

Since we do not yet advance the next measurement list entry we want to get,
we will at this point always request to receive the whole measurement list.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>