check_pcrs: match PCR if no mb_refstate is provided
If the values for mb_refstate are empty, the PCRs that are present in
the tpm_policy and belong to the measured boot process are ignored.

This patch check that the PCRs form the policy matches the ones that
comes from the quote when there is no measured boot data.

Fix #694

Signed-off-by: Alberto Planas <aplanas@suse.com>