agent: Get boottime and transfer to verifier
Add psutil as a dependency to be able to get the system's boottime and send
the boottime of the system an agent is running on to the verifier every
time we send a quote response. The verifier will then use the boottime to
start over with the attestation (start at entry 0) if it finds that the
boottime of the system has changed from the last time it did attestation.

We require psutil >=5.4.2 since Ubuntu Bionic provides this version and
RHEL 8 provides 5.4.3.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>