verifer: Fix signature and allowlist evaluation bahavior change
The recent move to ast has changed the behavior of the signature and
allow list evaluation. In particular it was now evaluating the allow list
after a successful signature verification even if the file was not in the
allow list. Modify the code to revert to the old behavior:

If there is also an allowlist verify the file against that but only do this if:
 - we did not evaluate the signature (valid_siganture = False)
 - the signature is valid and the file is also in the allowlist

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>