Automatically calculates the boot_aggregate from the measured boot log.
This PR uses the tsseventextend tool (found on package tss2), which
reads the binary measured boot log (sent from the `agent` to the
`verifier`), and then uses it to calculate the boot_aggregate.
This is done by using the (also calculated) values for PCRs 0-7(9)
also from the measured boot log (the need for contemplate both
PCRs 0-7 and 0-9 comes from the combination of older kernels and
older versions of grub).

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>