ima: Populate verifier keyrings with keys taken from ima-buf log line
Take the contents of the ima-buf log entry and try to determine whether
the data part is a key by having it parsed as a key.
Validate the ima-buf using the validator for ima-ng and if that is
accepted, add the given key to the verifiers virtual keyring given then
name of the keyring from the ima-buf filename part. This key will then
be used for signature verification.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>