ima: Extend alist with a map for 'keyrings' and convert from flat list
Extend the JSON alist with a map for 'keyrings' and approved hashes for
keys to be loaded into keyrings. Bump up the version of the supported
allowlist format.

If an allowlist has a version, accept any version <= current version
assuming that we provide backwards compatibility.

Convert entries from the (legacy) flat allowlist starting with %keyring:,
such as %keyring:.ima, to get entries in the JSON allow list. In this
example the .ima keyring would get a hash entry.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>