Add default verifier-check value
Set the verifier-check to True in do_cvreactivate and do_cvdelete,
according to the default value set for the corresponding arg set
in tenant.py.
This handles the issue in tenant_webapp.py where the do_cvreactivate
and the do_cvdelete functions are called without specifying the value
for verifier-check.

Also fix a typo and a reduntant parenthesis.

Signed-off-by: Riccardo Pittau <elfosardo@gmail.com>