ima: Return an empty allowlist rather than a plain empty list
read_allowlist() returns a plain empty list if al_path is None or ''.
However, we should return a dict following the new format since that is
what is returned when the list was converted from a JSON string or
converted from the old flat file format.

Create an empty_allowlist serving as a template to return in case an empty
list is to be returned and to extend when converting from the old flat
file.

Also, the code was reading the contents of the file into alist but this
alist was then never read from. Remove the unnecessary code and only
leave the 'test' for availability of the file.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>