Additional fix for the processing of "tpm_policy"
The "tpm_policy" dictionary (per-`agent` attribute) is converted
to a string by the code on `tenant.py` (used by `keylime_tenant`).
It is is then processed by the `verifier` even before it is stored
on the database, causing an error. This PR ensures that the the
contents of this attribute is selectively converted back to a
dictionary before processing.

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>