ima: Implement ima-buf validator and validate keys on keyrings (#725)
Implement the ima-buf validator and check whether the data part of the
ima-buf log entry is a keyring to check it against a keyring entry. Use the
'keyrings' selector for selecting the 'keyrings' map in the allowlist when
checking the keyring name using the ima-ng validator.

Extend the JSON alist with a map for 'keyrings' and approved hashes for
keys to be loaded into keyrings. Bump up the version of the supported
allowlist format.

If an allowlist has a version, accept any version <= current version
assuming that we provide backwards compatibility.

Convert entries from the (legacy) flat allowlist starting with %keyring:,
such as %keyring:.ima, to get entries in the JSON allow list. In this
example the .ima keyring would get a hash entry.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>