ci: build and publish container images
Four container images are build: keylime_base, keylime_verifier,
keylime_registrar and keylime_tenant.

In keylime_base Keylime is installed using installer.sh script.
All the other images are based on keylime_base and differ only in their
entrypoints and some configuration.

Docker tags are generated for each commit on the master branch and can be
referenced with master or sha-COMMIT_HASH. For git tags a tag with the
same name is created and the latest tag updated.

Images are currently published to ghcr.io

Signed-off-by: Thore Sommer <mail@thson.de>