agent: ensure that EK is in PEM format when used as uuid
The keylime.conf states that when hash_ek is used the uuid is the EK
public key in PEM format hashed with sha256. This was not the case instead
the internal tss format base64 encoded was hashed.

Now if the TPM has an EK certificate the uuid can be easily derived with:
tpm2_nvread 0x1c00002 | openssl x509 -inform DER -pubkey -noout | sha256sum

Signed-off-by: Thore Sommer <mail@thson.de>