Add tests for ca_impl_cfssl and ca_util
Following the replacement of M2Crypto with python-cryptography in
the previous commit, it looks like a good idea to improve our test
coverage of the use we make of that library.

This commit adds some tests for ca_util and also for the cfssl CA
implementation. Tests requiring cfssl will be skipped if it is not
found on the PATH.

Signed-off-by: Sergio Correia <scorreia@redhat.com>