ima: Extend allowlist to be able to handle generic ima-buf entries
Extend the allowlist to be able to handle generic ima-buf entries that for
example appear when one adds 'ima_policy=critical_data' to the Linux boot
command line parameters. With this policy one may get the following entries
in the measurement log:

10 6ddb524d7c630f0d64ac1c048ebf6e1a4b2099ac ima-buf sha256:acc2... kernel_version 352e31322e382d333030372e666333342e7838365f3634
10 b8562cfba9b8e9302f72b4cb9c3f4bbb07264589 ima-buf sha256:2dbb... selinux-policy-hash 5a5cd4968d0476b704e24f4ede5b372bda0b84fb46ccd2f53a7f93026218b08c

The kernel version's sha256 hash can be reproduced like this:

uname -r | tr -d '\n' | sha256sum -> acc23743...

The number string behind the kernel_version above is in human readable
format '5.12.8-3007.fc34.x86_64', so it's the same as
'uname -r | tr -d '\n''

The hash following the seclinux-policy-hash can be recreated using
sha256sum < /sys/fs/selinux/policy

The sha256 of the actual measurement is the sha256 of the binary
representation of the policy's hash.

So the simplest way to support this is to add an ima-buf entry to the
allowlist like shown below to be able to accept these measurements if they
show up:

{
    "meta" : {
       "version": 3
    },
    "release": 0,
    "hashes": {
        [...]
    },
    "ima-buf": {
        "kernel-version": ["acc23743f9ff73d197875292691fa3ac586940c2fb2bde9b0c692cc6c5cb5b0a"],
        "selinux-policy-hash": ["2dbb2f53d0196d9e2c11d55f61549cc828740d6d9428e7df0c354326f979ea27"]
    }
}

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>