ima: Use seek() to get to the last known last entry
Use seek() to get to the last known last entry in the IMA measurement
list. Track the size of the IMA measurement list two last known number
of entries in a singleton class that helps us seek to the last entry
and return the tail of the measurement list at the n-th entry as re-
quested by the caller.

We have to remember the last two known number of entries and file sizes
since the validator may not get a quote containing the last entry but
that of the previous last entry and will on the next attestation request
again ask for the same entry.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>