ima: Fix spurious attestation failures
Since IMA log appending and PCR extensions are not atomic we can get into
the situation where we get a PCR quote that does not contain the PCR 10
taking into account the latest measurement. This can happen if there has
been a single measurement added to the list. In this case we will not find
the matching PCR 10 when walking the measurement list but have to look for
a match before walking the measurement list.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>