scripts: Implement create_policy to create the JSON allowlist from files
Using either the local IMA ascii_runtime_measurements file or a file with
the same contents, extract keyring and critical data from it and add key
value pairs to their respective maps in the JSON-formatted allowlist.
Similarly, either read a flat-file allowlist and add its entries to the
JSON-formatted allowlist hashes map or read entries from the local IMA
ascii_runtime_measurement, or a file with the same contents, and add
filename - hash pairs to the hashes map.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>