elchecking: add new tests
EvSeperatorTest and EvEfiActionTest are test for common constant values
in the event log.

OnceTest is a test that can be only run once successfully. This is useful
for events that we only expect once in the log. It prohibits that this
test can be reused to validate other entries by changing the event type.

Signed-off-by: Thore Sommer <mail@thson.de>