config: move out 'convert' function
The convert function is used to decode containers that have byte arrays,
to strings.

This commit moves it to the cmd_exec module, and rename it to decode.

Also move out yaml_to_dict, that depends on it.  This function drops the
logger parameter, as is already declared in the module.

Signed-off-by: Alberto Planas <aplanas@suse.com>