ca: support back old cyptography API
After [1] we moved from M2Crypto to cryptography, and we started to drop
the use of the, now optional, `backend` parameter.  This is OK as we are
explicit on the minimal version of cryptography, but we still use in
other modules (like crypto.py or tpm_main.py for example) the old API.

This patch use the optional `backend` parameter to unify the API usage,
and still be compatible with some older versions of cryptography.

[1] https://github.com/keylime/keylime/pull/747

Signed-off-by: Alberto Planas <aplanas@suse.com>