config: move directory utilities to fs_util
Small refactoring where we move all the directory manipulation utilities
from the `config` module to the `fs_util` one.

Also drop chownroot function.

After a `os.makedirs` we are calling `chownroot` to set root as the
owner of the new directory.

But if we are currently under the `root` user, we should do not need
to call `chownroot` for the newly created directory.

Signed-off-by: Alberto Planas <aplanas@suse.com>