setup.py, config: install default configuration into package path
The keylime.conf is copied into the package path instead of trying to
install it globally into /etc which fails if the user is non root.

We will fallback on this configuration the other potential ones do not
exist.

We now also pass the parsed config directly to the logging config.

Signed-off-by: Thore Sommer <mail@thson.de>