revocation_notifier: kill ZeroMQ broker if it blocks more than 5s
device.destroy() is blocking and in some cases block for infinity.
To not hang on shutdown the process is killed after five seconds.

os.setpgrp() is used to ensure that we can manually shut down the process,
because otherwise the process would stop at the same time that the verifier
receives a SIGINT or SIGTERM.

Signed-off-by: Thore Sommer <mail@thson.de>