verifier: wait for connections to be closed before stopping ioloop
We now first stop the server to not accept new incoming connections and
then wait for all existing connections to be closed.

Instead of forking with tornado.process.fork_processes() we now use
Python's multiprocessing. This allows us to easily attach signal handlers
to the server processes. Note that we now do not restart a process if it
existed unexpectedly.

Signed-off-by: Thore Sommer <mail@thson.de>