revocation_notifier: convert the data to str in the notifiers
python-requests will internally use either simplejson (preferred)
or the built-in the json module; when it uses the latter, it may
encounter issues handling bytes instead of strings.

Let's convert the data to str beforand, to avoid this issue.

In the json module, __bytes_to_str() was renamed to bytes_to_str()
as it's now being used outside the module.

Signed-off-by: Sergio Correia <scorreia@redhat.com>