services: make the services run as keylime user instead of root
The services for the registrar and verifier are changed to run as keylime
user. The agent starts as root but the installer.sh changes the "run_as"
option to drop privileges to keylime:tss.

The installer.sh is changed to create the keylime user, create the correct
directories, change the permissions to the keylime user and to also install
the secure mount point.

Signed-off-by: Thore Sommer <mail@thson.de>